requirements-dev.txt
asgiref==3.2.10
certifi==2021.10.8
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1
django-binary-database-files==1.0.15
django-cors-headers==3.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
djoser==2.0.3
Pillow==9.1.1
PyJWT==2.3.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
static3==0.7.0
wincertstore==0.2
